Add firefox and twitterrific, separate in multiple files
