+++ Wire-iOS/Sources/UserInterface/Settings/Link Opening/MapOpening.swift
@@ -1,0 +1,68 @@
+//
+// Wire
+// Copyright (C) 2016 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+
+enum MapsOpeningOption: Int, LinkOpeningOption {
+
+    case apple, google
+
+    static var allOptions: [MapsOpeningOption] {
+        return [.apple, .google]
+    }
+
+    var displayString: String {
+        switch self {
+        case .apple: return "open_link.maps.option.apple".localized
+        case .google: return "open_link.maps.option.google".localized
+        }
+    }
+
+    var isAvailable: Bool {
+        switch self {
+        case .apple: return true
+        case .google: return UIApplication.shared.googleMapsInstalled
+        }
+    }
+
+}
+
+
+extension URL {
+
+    public func openAsLocation() -> Bool {
+        let saved = MapsOpeningOption(rawValue: Settings.shared().mapsLinkOpeningOptionRawValue) ?? .apple
+        switch saved {
+        case .apple: return false
+        case .google:
+            return UIApplication.shared.openURL(self)
+        }
+    }
+    
+}
+
+
+// MARK: - Private
+
+
+fileprivate extension UIApplication {
+
+    var googleMapsInstalled: Bool {
+        return canHandleScheme("comgooglemaps://")
+    }
+
+}
