+++ Wire-iOS/Wire-Info.plist
@@ -78,8 +78,10 @@
 	<array>
 		<string>comgooglemaps</string>
 		<string>tweetbot</string>
+		<string>twitterrific</string>
 		<string>googlechrome</string>
 		<string>googlechromes</string>
+		<string>firefox</string>
 		<string>org-appextension-feature-password-management</string>
 	</array>
 	<key>LSRequiresIPhoneOS</key>
